Use NativeLibrary.Load instead of dlopen
